+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The next release will require at least [Go 1.21].
 ### Added
 
 - Support [Go 1.22]. (#4890)
+- Support configuring a custom Proxy Function for the `go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp` HTTP Client Transport.
 
 ## [1.23.1] 2024-02-07